pyproject.toml
[project]
name = "projeto"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "dj-database-url (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "drf-spectacular-sidecar (>=2026.1.1,<2027.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-isort (>=7.0.0,<8.0.0)",
    "flake8-black (>=0.4.0,<0.5.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]
